Fix trap handling on successful exits

+++ release/tar/linux/opensearch-dashboards-tar-build.sh
@@ -113,7 +113,10 @@ OPENSEARCH_DASHBOARDS_PLUGINS_URLS=`yq eval '.products.opensearch-dashboards.plu
 
 TRAP_SIG_LIST="TERM INT EXIT"
 function delete_temp_folders() {
-    rm -rf $WORKING_DIR $PLUGINS_TEMP $TARGET_DIR
+    if [ $? -ne 0 ]; then
+      echo "Deleting the temp folders"
+      rm -rf $WORKING_DIR $PLUGINS_TEMP $TARGET_DIR
+    fi
 }
 trap delete_temp_folders $TRAP_SIG_LIST
